Vehicle: Move Generic Code to MAVLink Module
