+++ src/AutoPilotPlugins/APM/APMSensorsComponentController.h
@@ -10,7 +10,7 @@
 #pragma once
 
 #include "FactPanelController.h"
-#include "MAVLinkLib.h"
+#include "QGCMAVLink.h"
 
 #include <QtQuick/QQuickItem>
 #include <QtCore/QObject>
@@ -95,18 +95,6 @@ public:
     bool compassSetupNeeded (void) const;
     bool accelSetupNeeded   (void) const;
 
-    typedef enum {
-        CalTypeAccel,
-        CalTypeGyro,
-        CalTypeOnboardCompass,
-        CalTypeLevelHorizon,
-        CalTypeCompassMot,
-        CalTypePressure,
-        CalTypeAccelFast,
-        CalTypeNone
-    } CalType_t;
-    Q_ENUM(CalType_t)
-
     bool compass1CalSucceeded(void) const { return _rgCompassCalSucceeded[0]; }
     bool compass2CalSucceeded(void) const { return _rgCompassCalSucceeded[1]; }
     bool compass3CalSucceeded(void) const { return _rgCompassCalSucceeded[2]; }
@@ -125,7 +113,7 @@ signals:
     void resetStatusTextArea                    (void);
     void waitingForCancelChanged                (void);
     void setupNeededChanged                     (void);
-    void calibrationComplete                    (CalType_t calType);
+    void calibrationComplete                    (QGCMAVLink::CalibrationType calType);
     void compass1CalSucceededChanged            (bool compass1CalSucceeded);
     void compass2CalSucceededChanged            (bool compass2CalSucceeded);
     void compass3CalSucceededChanged            (bool compass3CalSucceeded);
@@ -172,7 +160,7 @@ private:
     
     bool _showOrientationCalArea;
     
-    CalType_t _calTypeInProgress;
+    QGCMAVLink::CalibrationType _calTypeInProgress;
 
     uint8_t _rgCompassCalProgress[3];
     bool    _rgCompassCalComplete[3];
