+++ src/AutoPilotPlugins/PX4/PowerComponentController.cc
@@ -19,7 +19,7 @@ void PowerComponentController::calibrateEsc(void)
 {
     _warningMessages.clear();
     connect(_vehicle, &Vehicle::textMessageReceived, this, &PowerComponentController::_handleVehicleTextMessage);
-    _vehicle->startCalibration(Vehicle::CalibrationEsc);
+    _vehicle->startCalibration(QGCMAVLink::CalibrationEsc);
 }
 
 void PowerComponentController::startBusConfigureActuators(void)
