+++ src/MAVLink/MAVLinkFTP.h
@@ -10,8 +10,11 @@
 #pragma once
 
 #include <QtCore/QString>
+#include <QtCore/QLoggingCategory>
 
 #include "MAVLinkLib.h"
+
+Q_DECLARE_LOGGING_CATEGORY(MavlinkFTPLog)
 
 class MavlinkFTP {
 public:
