+++ src/MAVLink/MAVLinkStreamConfig.h
@@ -10,6 +10,9 @@
 #pragma once
 
 #include <QtCore/QVector>
+#include <QtCore/QLoggingCategory>
+
+Q_DECLARE_LOGGING_CATEGORY(MAVLinkStreamConfigLog)
 
 /**
  * @class MAVLinkStreamConfig
