+++ src/MAVLink/SysStatusSensorInfo.h
@@ -13,6 +13,9 @@
 
 #include <QtCore/QMap>
 #include <QtCore/QObject>
+#include <QtCore/QLoggingCategory>
+
+Q_DECLARE_LOGGING_CATEGORY(SysStatusSensorInfoLog)
 
 /// Class which represents sensor info from the SYS_STATUS mavlink message
 class SysStatusSensorInfo : public QObject
